backend/ai_services/views.py
"""
AI Services Views
"""
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from django.conf import settings
import google.generativeai as genai
import json
import logging

logger = logging.getLogger(__name__)

# Configure Gemini AI
if settings.GEMINI_API_KEY:
    genai.configure(api_key=settings.GEMINI_API_KEY)


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def generate_clarifying_questions(request):
    """Generate clarifying questions for patient data"""
    try:
        patient_data = request.data.get('patient_data', {})
        
        if not patient_data:
            return Response({
                'success': False,
                'error': {
                    'code': status.HTTP_400_BAD_REQUEST,
                    'message': 'Bemor ma\'lumotlari kiritilmagan'
                }
            }, status=status.HTTP_400_BAD_REQUEST)
        
        # TODO: Implement Gemini API call
        # For now, return mock data
        questions = [
            "Qachondan boshlab shikoyatlar paydo bo'ldi?",
            "Shikoyatlar doimiy yoki davriy?",
            "Qanday holatlarda shikoyatlar kuchayadi?",
        ]
        
        return Response({
            'success': True,
            'data': questions
        })
    
    except Exception as e:
        logger.error(f"Error generating clarifying questions: {e}")
        return Response({
            'success': False,
            'error': {
                'code': status.HTTP_500_INTERNAL_SERVER_ERROR,
                'message': 'Savollar yaratishda xatolik yuz berdi'
            }
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def recommend_specialists(request):
    """Recommend AI specialists based on patient data"""
    try:
        patient_data = request.data.get('patient_data', {})
        
        if not patient_data:
            return Response({
                'success': False,
                'error': {
                    'code': status.HTTP_400_BAD_REQUEST,
                    'message': 'Bemor ma\'lumotlari kiritilmagan'
                }
            }, status=status.HTTP_400_BAD_REQUEST)
        
        # TODO: Implement Gemini API call
        recommendations = [
            {'model': 'Cardiologist', 'reason': 'Yurak-qon tomir tizimi bilan bog\'liq shikoyatlar'},
            {'model': 'Neurologist', 'reason': 'Nevrologik simptomlar mavjud'},
        ]
        
        return Response({
            'success': True,
            'data': {'recommendations': recommendations}
        })
    
    except Exception as e:
        logger.error(f"Error recommending specialists: {e}")
        return Response({
            'success': False,
            'error': {
                'code': status.HTTP_500_INTERNAL_SERVER_ERROR,
                'message': 'Mutaxassislar tavsiya qilishda xatolik yuz berdi'
            }
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def generate_diagnoses(request):
    """Generate initial differential diagnoses"""
    try:
        patient_data = request.data.get('patient_data', {})
        
        if not patient_data:
            return Response({
                'success': False,
                'error': {
                    'code': status.HTTP_400_BAD_REQUEST,
                    'message': 'Bemor ma\'lumotlari kiritilmagan'
                }
            }, status=status.HTTP_400_BAD_REQUEST)
        
        # TODO: Implement Gemini API call
        diagnoses = [
            {
                'name': 'O\'tkir miokard infarkti',
                'probability': 75,
                'justification': 'Ko\'rsatkichlar va simptomlar asosida',
                'evidenceLevel': 'High'
            }
        ]
        
        return Response({
            'success': True,
            'data': diagnoses
        })
    
    except Exception as e:
        logger.error(f"Error generating diagnoses: {e}")
        return Response({
            'success': False,
            'error': {
                'code': status.HTTP_500_INTERNAL_SERVER_ERROR,
                'message': 'Tashxis yaratishda xatolik yuz berdi'
            }
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def run_council_debate(request):
    """Run AI council debate"""
    try:
        patient_data = request.data.get('patient_data', {})
        diagnoses = request.data.get('diagnoses', [])
        specialists = request.data.get('specialists', [])
        
        if not patient_data or not diagnoses:
            return Response({
                'success': False,
                'error': {
                    'code': status.HTTP_400_BAD_REQUEST,
                    'message': 'Kerakli ma\'lumotlar kiritilmagan'
                }
            }, status=status.HTTP_400_BAD_REQUEST)
        
        # TODO: Implement full council debate logic
        # This would be a complex async operation
        
        return Response({
            'success': True,
            'message': 'Konsilium munozarasi boshlandi',
            'data': {
                'status': 'processing',
                'message': 'Bu uzoq vaqt olishi mumkin'
            }
        })
    
    except Exception as e:
        logger.error(f"Error running council debate: {e}")
        return Response({
            'success': False,
            'error': {
                'code': status.HTTP_500_INTERNAL_SERVER_ERROR,
                'message': 'Konsilium munozarasi boshlamasda xatolik yuz berdi'
            }
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
