backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-yasg==1.21.7
whitenoise==6.6.0
gunicorn==21.2.0
requests==2.31.0
django-redis==5.4.0
